printf: Fix dynamic string args.
Trusts that the frontend emits the literal strings to
the global address space. If we emitted them to the constant
space, there would be illegal ascasts between constant and
generic.

dynamic_str_args is a test suite that passes with PoCL/CPU
but fails with LZ. We need a conditional XFAIL checking at
runtime if we want to add it to the test suite.